Unite GradientEdit and GradientEditor as editor-only widget
